working frontend and html without two main controllers
