--- budget/src/main/java/com/prospero/budget/Home.java
@@ -1,5 +1,0 @@
-package com.prospero.budget;
-
-public class Home {
-    private String username;
-}
