+++ budget/src/main/resources/templates/budgeter.html
@@ -23,7 +23,7 @@
     <nav>
         <button onclick="window.location.href='/'">🏠<br>Dashboard</button>
         <button onclick="window.location.href='/budgeter'">💰<br>Budgets</button>
-        <button onclick="window.location.href='/invest.html'">💳<br>Investing</button>
+        <button onclick="window.location.href='/invest'">💳<br>Investing</button>
      </nav>
 
      <main>
