+++ budget/src/main/resources/templates/index.html
@@ -24,9 +24,9 @@
 
     <!-- Side Nav -->
     <nav>
-       <button onclick="window.location.href='/'">🏠<br>Dashboard</button>
-       <button onclick="window.location.href='/budgeter.html'">💰<br>Budgets</button>
-       <button onclick="window.location.href='/invest.html'">💳<br>Investing</button>
+      <button onclick="window.location.href='/'">🏠<br>Dashboard</button>
+      <button onclick="window.location.href='/budgeter'">💰<br>Budgets</button>
+      <button onclick="window.location.href='/invest'">💳<br>Investing</button>
     </nav>
 
     <!-- Main Content -->
