+++ budget/src/main/resources/templates/invest.html
@@ -25,8 +25,8 @@
     <!-- Side Nav -->
     <nav>
        <button onclick="window.location.href='/'">🏠<br>Dashboard</button>
-       <button onclick="window.location.href='/budgeter.html'">💰<br>Budgets</button>
-       <button onclick="window.location.href='/invest.html'">💳<br>Investing</button>
+       <button onclick="window.location.href='/budgeter'">💰<br>Budgets</button>
+       <button onclick="window.location.href='/invest'">💳<br>Investing</button>
     </nav>
 
     <!-- Main Content -->
